chore: adjust the file directory structure
