+++ packages/ui/src/components/Node/ItemCollection.test.ts
@@ -1,7 +1,6 @@
 import { describe, it, expect } from 'vitest';
 import { ItemCollection } from './ItemCollection';
-import { nested } from './itemExamples/nested';
-import { normal } from './itemExamples/normal';
+import { nested, normal, threeTierNested } from './mock';
 
 describe('toTable', () => {
   it('should correctly extract headers and rows from normal data', () => {
@@ -29,32 +28,6 @@ describe('toTable', () => {
   });
 
   it('should correctly extract headers and rows from 3 nested Object', () => {
-    const jsonData = [
-      {
-        'foo1': 'bar1',
-        'foo2': 'bar2',
-        'foo3': 'bar3'
-      },
-      {
-        'foo1': 'bar1',
-        'foo2': 'bar2',
-        'foo3': 'bar3'
-      },
-      {
-        'foo1': 'bar1',
-        'foo2': 'bar2',
-        'foo3': {
-          'foo1': 'bar1',
-          'foo2': 'bar2',
-          'foo3': {
-            'foo1': 'bar1',
-            'foo2': 'bar2',
-            'foo3': 'bar3'
-          },
-        },
-      }
-    ];
-
     const expectedHeaders = ['foo1', 'foo2', 'foo3', 'foo3.foo1', 'foo3.foo2', 'foo3.foo3.foo1', 'foo3.foo3.foo2', 'foo3.foo3.foo3'];
     const expectedContent = [
       ['bar1', 'bar2', 'bar3', undefined, undefined, undefined, undefined, undefined ],
@@ -62,7 +35,7 @@ describe('toTable', () => {
       ['bar1', 'bar2', undefined, 'bar1', 'bar2', 'bar1', 'bar2', 'bar3'],
     ];
 
-    const { headers, rows } = new ItemCollection(jsonData).toTable();
+    const { headers, rows } = new ItemCollection(threeTierNested).toTable();
 
     expect(headers).toEqual(expectedHeaders);
     expect(rows).toEqual(expectedContent);
