--- packages/ui/src/components/Node/itemExamples/createLargeColsFn.ts
@@ -1,5 +1,0 @@
-export const createLargeColsFn = (count: number) => Array.from({ length: count }, (_, index) => ({
-  [`column_${index}`]: `Value ${index}`,
-})).reduce((acc, item) => {
-  return { ...acc, ...item };
-});
