--- packages/ui/src/components/Node/itemExamples/normal.ts
@@ -1,11 +1,0 @@
-export const normal = {
-  property_a: 10000,
-  property_b: 20000,
-  property_c: 30000,
-  property_d: 40000,
-  property_e: 50000,
-  property_f: 60000,
-  property_g: 70000,
-  property_h: 80000,
-  property_i: 90000,
-}
