--- packages/ui/src/components/Node/itemExamples/oversize.ts
@@ -1,6 +1,0 @@
-export const oversize = {
-  normal: 'normal',
-  long_property: 'long_'.repeat(10000), // Long value
-  normal2: 'normal',
-  ['long_'.repeat(20)]: 'normal', // Long key
-}
